refactor: modularize frontend scripts
