+++ js/state.js
@@ -1,0 +1,24 @@
+export const projectData = {
+  title: "Projeto LoreLoom",
+  content: "",
+  characters: [],
+  locations: [],
+  items: [],
+  languages: [],
+  timeline: [],
+  notes: [],
+  economy: {
+    currencies: [],
+    resources: [],
+    markets: []
+  }
+};
+
+export const editingIds = {
+  character: null,
+  location: null,
+  item: null,
+  language: null,
+  event: null,
+  note: null
+};
