+++ loreloom.html
@@ -865,6 +865,6 @@
       </div>
     </div>
   </div>
-  <script src="js/scripts.js"></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
